+++ taptest/errors_test.lua
@@ -7,11 +7,13 @@ local cluster = require('cluster')
 
 local test = tap.test('cluster.cfg')
 
-test:plan(8)
+test:plan(12)
 
 local function check_error(expected_error, fn, ...)
     local ok, err = fn(...)
-    test:diag('%s', err)
+    for _, l in pairs(string.split(tostring(err), '\n')) do
+        test:diag('%s', l)
+    end
     test:like(err.err, expected_error, expected_error)
 end
 
@@ -24,7 +26,7 @@ check_error('Can not create workdir "/dev/null"',
 
 check_error('Missing port in advertise_uri "localhost"',
     cluster.cfg, {
-        workdir = '.',
+        workdir = './dev',
         advertise_uri = 'localhost',
     }
 )
@@ -33,7 +35,7 @@ local _sock = socket('AF_INET', 'SOCK_DGRAM', 'udp')
 local ok = _sock:bind('0.0.0.0', 3301)
 check_error('Socket bind error',
     cluster.cfg, {
-        workdir = '.',
+        workdir = './dev',
         advertise_uri = 'localhost:3301',
     }
 )
@@ -42,27 +44,72 @@ _sock = nil
 
 check_error('Can not ping myself: ping was not sent',
     cluster.cfg, {
-        workdir = '.',
+        workdir = './dev',
         advertise_uri = 'invalid-host:33004',
     }
 )
 
-check_error([[module 'unknown' not found]],
+check_error([[module 'unknown%-role' not found]],
     cluster.cfg, {
-        workdir = '.',
+        workdir = './dev',
         advertise_uri = 'localhost:33003',
-        roles = {'unknown'},
+        roles = {'unknown-role'},
     }
 )
 
-package.preload['mymodule'] = function()
-    error('My module can not be loaded')
-end
-check_error('My module can not be loaded',
+check_error([[module 'unknown%-auth' not found]],
     cluster.cfg, {
-        workdir = '.',
+        workdir = './dev',
+        advertise_uri = 'localhost:33005',
+        auth_backend_name = 'unknown-auth',
+    }
+)
+
+package.preload['myrole'] = function()
+    error('My role can not be loaded')
+end
+check_error('My role can not be loaded',
+    cluster.cfg, {
+        workdir = './dev',
         advertise_uri = 'localhost:33002',
-        roles = {'mymodule'},
+        roles = {'myrole'},
+    }
+)
+
+package.preload['myauth'] = function()
+    error('My auth can not be loaded')
+end
+check_error('My auth can not be loaded',
+    cluster.cfg, {
+        workdir = './dev',
+        advertise_uri = 'localhost:33006',
+        auth_backend_name = 'myauth',
+    }
+)
+
+package.preload['auth-unknown-method'] = function()
+    return {
+        unknown_method = function() end,
+    }
+end
+check_error('unexpected argument callbacks.unknown_method to set_callbacks',
+    cluster.cfg, {
+        workdir = './dev',
+        advertise_uri = 'localhost:33007',
+        auth_backend_name = 'auth-unknown-method',
+    }
+)
+
+package.preload['auth-invalid-method'] = function()
+    return {
+        check_password = 'not-a-function',
+    }
+end
+check_error('bad argument callbacks.check_password to set_callbacks',
+    cluster.cfg, {
+        workdir = './dev',
+        advertise_uri = 'localhost:33008',
+        auth_backend_name = 'auth-invalid-method',
     }
 )